+++ VimR/VimR/FoundationCommons.swift
@@ -7,5 +7,10 @@ import Foundation
 import os
 
 extension URL {
-  var direntType: UInt8 { (self as NSURL).direntType() }
+  var direntType: Int16 {
+    if self.isRegularFile { return Int16(DT_REG)}
+    if self.hasDirectoryPath { return Int16(DT_DIR) }
+
+    return Int16(DT_UNKNOWN)
+  }
 }