+++ VimR/VimR/ScoredUrl.h
@@ -9,15 +9,12 @@
 @interface ScoredUrl : NSObject
 
 @property (readonly, nonnull) NSURL *url;
-@property (readonly) NSInteger score;
+@property (readonly) double score;
 
+- (instancetype _Nonnull)initWithUrl:(NSURL * _Nonnull)url score:(double)score;
+- (NSString * _Nonnull)description;
 - (BOOL)isEqual:(id _Nullable)other;
 - (BOOL)isEqualToUrl:(ScoredUrl * _Nullable)url;
 - (NSUInteger)hash;
-
-- (instancetype _Nonnull)initWithUrl:(NSURL * _Nonnull)url score:(NSInteger)score;
-- (NSString * _Nonnull)description;
-+ (instancetype _Nonnull)urlWithUrl:(NSURL * _Nonnull)url score:(NSInteger)score;
-
 
 @end